Cleanup utility function names
